# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/businesses/stats": {
            "get": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::stats",
                "parameters": [
                    {
                        "name": "top_left",
                        "in": "query",
                        "description": "Top Left of location (GPS)",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "bottom_right",
                        "in": "query",
                        "description": "Bottom right of location (GPS)",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats data"
                    }
                }
            }
        },
        "/api/v1/businesses": {
            "get": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::index",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Lat location (GPS)",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Lng of location (GPS)",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Lng of location (GPS)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "map_preset_id",
                        "in": "query",
                        "description": "Map Preset ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List businesses"
                    }
                }
            },
            "put": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "uuid of business",
                        "required": true,
                        "schema": {
                            "type": "char(36)"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of business",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Lat",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Lng",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business updated"
                    },
                    "400": {
                        "description": "Business not found"
                    }
                }
            },
            "post": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of business",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Lat",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Lng",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "BusinessResource"
                    }
                }
            },
            "delete": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "uuid of business to delete",
                        "required": true,
                        "schema": {
                            "type": "char(36)"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business updated"
                    },
                    "400": {
                        "description": "Business not found"
                    }
                }
            }
        },
...
        "/api/v1/businesses/{id}/avatar": {
            "post": {
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::updateAvatar",
                "parameters": [
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "Avatar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload successfully"
                    },
                    "422": {
                        "description": "Cannot upload avatar"
                    }
                }
            }
        },
...
    }
}